Use secrets for gpg keys and passphrase
