+++ containers/base_rpm.go
@@ -21,23 +21,23 @@ func RPMContainer(d *dagger.Client, opts *GPGOpts) *dagger.Container {
 	if !opts.Sign {
 		return container
 	}
+	gpgPublicKeySecret := d.SetSecret("gpg-public-key", opts.GPGPublicKey)
+	gpgPrivateKeySecret := d.SetSecret("gpg-private-key", opts.GPGPrivateKey)
+	gpgPassphraseSecret := d.SetSecret("gpg-passphrase", opts.GPGPassphrase)
 	return container.
+		WithSecretVariable("GPG_PUBLIC_KEY", gpgPublicKeySecret).
+		WithSecretVariable("GPG_PRIVATE_KEY", gpgPrivateKeySecret).
+		WithSecretVariable("GPG_PASSPHRASE", gpgPassphraseSecret).
 		WithExec([]string{"apt-get", "install", "-yq", "gnupg2"}).
+		WithExec([]string{"mkdir", "-p", "/root/.rpmdb/privkeys"}).
+		WithExec([]string{"mkdir", "-p", "/root/.rpmdb/passkeys"}).
+		WithExec([]string{"mkdir", "-p", "/root/.rpmdb/pubkeys"}).
+		WithExec([]string{"/bin/sh", "-c", "echo \"$GPG_PRIVATE_KEY\" | base64 -d > /root/.rpmdb/privkeys/grafana.key"}).
+		WithExec([]string{"/bin/sh", "-c", "echo \"$GPG_PASSPHRASE\" | base64 -d > /root/.rpmdb/passkeys/grafana.key"}).
+		WithExec([]string{"/bin/sh", "-c", "echo \"$GPG_PUBLIC_KEY\" | base64 -d > /root/.rpmdb/pubkeys/grafana.key"}).
 		WithNewFile("/root/.rpmmacros", dagger.ContainerWithNewFileOpts{
 			Permissions: 0400,
 			Contents:    RPMMacros,
-		}).
-		WithNewFile("/root/.rpmdb/privkeys/grafana.key", dagger.ContainerWithNewFileOpts{
-			Permissions: 0400,
-			Contents:    opts.GPGPrivateKey,
-		}).
-		WithNewFile("/root/.rpmdb/pubkeys/grafana.key", dagger.ContainerWithNewFileOpts{
-			Permissions: 0400,
-			Contents:    opts.GPGPublicKey,
-		}).
-		WithNewFile("/root/.rpmdb/passkeys/grafana.key", dagger.ContainerWithNewFileOpts{
-			Permissions: 0400,
-			Contents:    opts.GPGPassphrase,
 		}).
 		WithExec([]string{"gpg", "--batch", "--yes", "--no-tty", "--allow-secret-key-import", "--import", "/root/.rpmdb/privkeys/grafana.key"})
 }
